feat: Generate a password to be used for new users
The password generated can be given to users to login to their account
for the first time. Since we've set `must_change_password = true` this
is okay to share. The password is marked as sensitive so it won't be
stored in any logs and must be explicitly viewed by those managing
Terraform.