feat!: Update Module to Use YAML Configuration (#11)
* chore: Remove outdated examples

* chore: Remove non-standard tooling configuration

* chore: Bump versions

* chore: Prep for YAML configuration

* chore: Update gitignore for Terraform project

* chore: Add spec.yml to .gitignore

The spec.yml file is passed in as a variable to the module but should
not be part of the module itself.

* chore: Prep locals.tf for YAML configuration

* chore: Revert .gitignore

* feat: Populate the snowflake_user resources with configuration found in spec.yml

* chore: Cleanup unused Terraform files

* terraform-docs: automated action

* fix: Add token to checkout to triggers checks after automated commit

* feat: Generate a password to be used for new users

The password generated can be given to users to login to their account
for the first time. Since we've set `must_change_password = true` this
is okay to share. The password is marked as sensitive so it won't be
stored in any logs and must be explicitly viewed by those managing
Terraform.

* terraform-docs: automated action

* docs: Update README with spec.yml details

* docs: Update README to explain depends_on meta argument

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>